Update README.md
# WebhookDiscordDeleter-Spammer
A Webhook Deleter o/ Spammer. (image webhook changer BROKEN🥲)❌
# Version
The current version is 3.0, 1.0 and 2.0 is Non-Released.
# Date released
Date released in 23/02/2025
# Safe?
It is safe and legit
# Is it a virus?
You can check the source code if theres anything connected to a webhook or just read it and ask ChatGPT if you don't know what it means.
# If you need any help.
Please add discord (f1or1zf3nzrblx)
and show me whats the problem.